Fix
Now we get border colorsplits as expected! The big mistake was allowing interrupts to fire, which can happen between us writing byte 1 and 2 to one of the VDP ports, causing the internal byte flip flop to get out of sync. Disabling interrupts after writing characters fixes this (we don't want to do it before calling any BIOS routines, as they may enable ints again, which appears to be the case here actually). Also make sure border is set to color 0 before changing that color in the palette, and voila! Works great. :)