Spoof msx-dos
So many existing demos will actually ship as msx-dos disks with the demo alongside MSXDOS.SYS and COMMAND.COM, with an AUTOEXEC.BAT to start the demo when msx-dos boots. The major advantage here is that the msx-dos environment has a full 64k of mem banked in while running, so the demo can just use the full address space and not worry about doing a RAM search/mappers/etc. It also allows for use of msx-dos routines.

License-wise, this appears to be OK for msx-dos2, but it's unclear whether or not this is OK for msx-dos. This is a problem in my case, since msx-dos2 requires 128kb RAM to run, but my machine (sony hb-f1xd) only has 64kb RAM. I actually suspect this is the reason why many MSX2 demos require 128kb RAM to run.

However, after looking at the boot process, I noticed a couple fun things:
 - It's not msx-dos that ensures RAM is visible. It's unclear which part of the system is responsible for this, but this environment is actually initialized _before_ msx-dos is loaded.
 - Given a disk with a default msx-dos bootsector, the msx machine will boot msx-dos by loading a file called MSXDOS.SYS into memory at 0x100, and then transferring execution to it.

Here's the fun bit: nothing stops us from making our _own_ MSXDOS.SYS file, which can contain anything we want. Since there are multiple versions of msx-dos, there doesn't appear to be any checksums or hashes or anything for the file itself. So, if we produce our demo as a program called MSXDOS.SYS, the system will set up our full 64kb RAM visibility and load our demo automatically to a deterministic load addr, at which point we have full control of the system. And it doesn't even require a custom bootloader :) .

So yeah, that's what this patch does. Pretty neat! The only downside is now we don't have access to msx-dos' file loading routines, but since those appear to rely on the machine's internal bios/disk rom's anyways, it shouldn't make much difference. This keeps us from having to play nice with any sort of OS as well, and get the full mem range to ourselves.