Playing with BDOS system calls
Just testing the waters with some console output calls. Interestingly, it seems CR/LF doesn't seem to do anything when using CONOUT, which is a bit unexpected, but whatever. Would be interesting to see if I can load the print message from disk, and to see if interrupts still occur when reading from disk.