Test disk i/o via BDOS
Looks like it works! Currently able to read a message from a file on disk and display it with other BDOS routines.

From the folks at #msxdev it would appear interrupts are disabled for long periods of time (in the order of several tenths of a second) so we probably can't rely on that for loading dynamically in the background, unless we go _really_ low level with disk i/o, which I'm not that comfortable doing when we're talking about hw that varies as much as msx2 does.

I see basically two ways around this:
 - 1. Don't do any dynamic disk I/O during the demo. This would likely rely on storing compressed data for a bunch of parts in mem (possibly even using the second VRAM bank as extra storage here) and decrunching in the BG. This is actually quite similar to what C64 demos are doing as they will typically compress to reduce serial port bandwidth anyways, but instead of serial I/O we'd rely on DMA from VRAM, which we'd have a lot more control over timing-wise.
 - 2. Run some tests on carefully-prepared disk data. It may be possible that a single file arranged carefully into a specific record structure could remain open at all times during the demo, and small records could be loaded dynamically throughout in the bg at specific times where the timing interruptions would be tolerable. I have no idea if this would work, and I'd need to set up a test program that plays music/does some simple one-frame effect to test properly. This could be quite cool if it works.

In any case, this simple test was a success, so yay :)